Do `allItems` awaits in main items config view model instead of submenu models.
The few frames saved by propagating the async task (instead of its result) are mooted by the fact that none of these items can be seen on the first/default page of the mod config - so instead of loading the menu faster, it actually loads somewhat slower (since more assets need to be loaded, more contention on threads, etc.) and increases complexity needlessly.

It's better to have the items added to the UI a few frames later if users won't see the transition anyway.