Refactor: Extract common search and pager features.
Pagers, used to animate page transitions, behave more or less the same in each instance (config menu, mod item pages, soon to be quick action type pages). We can get rid of all the hacks originally done in `ConfigurationViewModel` and its view and replace it with this more refined version.

Item search is also a shared concern with quick slot config, so this was refactored at the same time.