Initial/partial implementation of quick slot editor.
Item template is cribbed largely from Pen Pals. Uses the now-shared pagination model to choose different item sources, for which the first two (inventory and item search) are implemented. Mod Actions still needs to be done.

The UI tries to avoid having any scrolling, because scrolling is confusing and hard to get right when there is supposed to be a clip region "above" the scroll bar (which occludes the scroll bar), and in addition, we end up with a bizarre amount of white space, which - while it also occurs without scrolling - is more obvious as to _why_ it occurs without scrolling and doesn't look as out-of-place.