Add checkboxes for quick slot confirmation and secondary action.
These are awkward to place because of the menu's layout, but underneath the (fixed-size, often with a lot of blank space) item selector seems like the least bad option.

Making this coherent requires removing the auto-close behavior, and instead ensuring that the preview updates. Otherwise the flow is awkward and users would have to remember to use the checkboxes first, or select the item and then reopen it to go back to the checkboxes. After the change, it works similar to the mod menu item editor where changes are realtime and the menu is closed by cancel/menu button.

Includes a few other trivial UI tweaks.