Implement most of quick slot behavior.
Slots are now activated and button suppression behavior seems to feel natural. For quick slots, we suppress the delay, otherwise it's not really "quick", and also force the trigger because otherwise the menu would reopen instantly. Still unsure if delay suppression is the right thing, but it felt less awkward than the alternative.

Unusable items are now drawn semi-transparent along with their prompts. The correct image is shown for the item that would actually match (e.g. Iridium Pickaxe for just a Pickaxe binding). In general, the logic should work for always trying to pick an approximate match with the best upgrade, given that there will be many cases without a perfect answer.

A few tasks remain but the system is largely functional.