Implement fallback, harmonyless integration between GMCM and new config.
Removes ALL legacy configuration classes and GMCM UI hackery, retargets GMCM sync to new configuration, and generally moves everything over to the new system.

We still need the Harmony patch version to make this properly seamless, but the fallback version should function even if Harmony fails in a future version, and it's good for testing in the meantime.

Some little bugs discovered with `TitleMenu` submenus, added as comments, need to be fixed soon.