Add preview to Styles config page.
Implementation uses a `RenderTarget2D` for the menu renderer, not because of performance (though it may be a small optimization) but because it is simply not possible to get the `BasicEffect` to draw in the foreground during UI rendering. It works in game, but not in menus. The workaround is to draw the entire menu to a texture and then redraw the texture (at unity scale, to avoid additional artifacts).