Refactor: Move GMCM sync to external "extension" assembly.
This lets us finally eliminate the hard GMCM dependency in the main mod, and offload the work to a dynamically-loaded assembly that can independently fail gracefully without causing the whole mod to be flagged as incompatible due to an internal change.

Harmony patching will also go here, since it is highly version dependent.