Add GMCM direct menu integration.
Will launch the custom menu directly from the GMCM's main page, instead of having to go through the intermediate GMCM page and click a button. The "manual" GMCM page still exists and will continue to exist as a fallback should the Harmony route fail.

Removes the old version of the GMCM DLL used for publicizing and fixes the hint path to use the version in the game mods folder.