Add activation methods back to input configuration, and in UI.
Reuses the old (and unused after refactoring) enum with some minor tweaks. The decision is to make the activation method _in addition_ to the bound button, because the keybinds can be removed. This also makes the configuration more coherent (somewhat) as any combination of the two settings, per action (primary vs. secondary), is valid, as long as primary and secondary don't conflict.

Includes only the configuration changes, not the input behavior to implement them yet.