Add opt-in right-stick chatbox suppression.
This is an important feature when using stick-based activation, either as a direct keybind or as secondary activation; the chatbox gets in the way.

Because the way chatbox activation is implemented makes it difficult to selectively suppress (i.e. only when actually using the menu), and because it is generally annoying in single-player games anyway, the option to disable it is a separate toggle and not automatic.

Requires Harmony dependency, and fixes some issues with Harmony logging.

Includes minor UI tweak: hide the "reopen on hold" option completely when not in hold mode, since moving the option to Advanced settings makes the relationship to Toggle Mode less obvious.