Check focused menu item even when index hasn't changed.
The reason for doing this is because `NextPage` and `PreviousPage` may keep the same index, i.e. if the thumbstick is held in place while using shoulder buttons to change the page. This changes the item, but not the index, so we need to check both, otherwise the activated item will not be consistent with the focused/drawn item.