Allow quick slots to be activated without the radial menu open.
This is always opt-in (configurable in quick slot popup editor), since every single controller button is technically mapped in vanilla and therefore we shouldn't ever force this behavior on users. It is also disabled for the A and X buttons to prevent what might appear to be an unrecoverable situation.

Also eliminates the `UpdateTicked` loop and moves everything into `UpdateTicking`, because `UpdateTicked` runs _after_ vanilla inputs get handled. We still need special suppressions for the trigger buttons, but the regular suppressions also need to run earlier to avoid leaking inputs rebound to quick slots.