Treat suppressed buttons as down for keeping menu open.
We need to do this for non-trigger buttons, otherwise the suppression will prevent `IsDown` from seeing it as down the next frame, and menus will close immediately.

To prevent accidental triggering, only do this when the menu is already active and the mode is set to "Hold". This should avoid situations where the suppression comes from another mod or some cause unrelated to the menu itself.