Add short trigger suppression on item activation, and allow re-triggering during non-interactive periods.
The trigger suppression (forced button suppression) helps prevent button presses from leaking into the "next" UI - for example, opening the SC settings using the right-trigger menu would previously cause that menu to advance to the next page 3-4 times out of 5, since the right trigger performs that action.

Re-triggering is done by checking for the button release during non-interactive state, which occurs when activation is being delayed in addition to the post-activation frames that apply to the previous change, and handling that as a separate state transition.

Both require a more extensive "pre-update" loop that runs regardless of whether the menu is actually enabled (player is ready/can move). To mitigate issues we try to do as little work as possible here, and ensure that the SMAPI bypass input suppression only happens during the interactive pre-update.