Add standard library items.
These are mostly easy to access other ways, aside from Mailbox, e.g. main menu is a single button press, map is the M key, etc. The main reason for creating them is for remapping the in-world keys; e.g. if the Select/Back button will be remapped to the remapping menu, then there needs to be another way to open the journal.

Makes several properties of `ModMenuItem` lazy, so that we can use lazy sprites and not try to read the texture too early, which can crash or be incorrect with content packs.