Define an enum for item activation type.
The old enum was changed to a `boolean` because the old enum was too narrow and didn't make sense in some cases; however, a boolean is going to be inflexible for future updates.

Switching to enum here also allows us to define an `Instant` type that will be used for button remappings, as a kind of suped-up primary.