Initial implementation of single-button (instant) tool use.
This mostly works - one-shot tools (pickaxe), melee weapons (swords) and chargeable tools (hoe) are all tested. Fishing Rod is still misbehaving, is at a point where it is timable but tries to immediately reel as soon as it hits the water. However, the remapping works well enough to allow the fish to be reeled with the remapped button as long as it is _cast_ with X... so, progress.

Doing this requires Harmony patching that's a little more invasive due to the amount of hardcoding of X button in the game. Also requires some additions to the Item API to represent the transient states, though most if not all API clients should be able to ignore them.