Revert the "always" on setting for quick slots.
This has now been redone and consolidated into the remapping system, so the half-baked quick slot version is obsolete and conflicting.

Not a straight git revert because other changes have been made, and most are worth keeping - i.e. the changes to the update timing, variable width of quick slot menu footer, and mod/project rename. Only the specific setting and associated UI needs to be reverted.