Allow remapping menu to be closed with the same hotkey used to open it.
This is usually something that works in games that have such menus, and feels awkward when it doesn't and a different key (start/B) needs to be pressed instead.

Implicitly ignores trigger buttons since they're required by the menu, and explicitly ignores d-pad buttons used to activate the menu since they're generally used for menu navigation (even if the left stick also works) and it would be even more awkward to have one d-pad direction close the menu.