Skip explicit `FishingRod` deactivation on release and allow simulate buttons _without_ reactivation during tool use.
The latter enables using the bobber bar and possibly other tool functions while the tool is active, which wouldn't normally be allowed in `Context.IsPlayerFree`. Actually reactivating is bad and causes all sorts of glitches (minigame jumps around, restarts fishing/reeling, etc.) but we can still apply the Harmony patch to trick the minigames into thinking the tool button is pressed.

To complete the circuit, needs a transpiler for `oldPadState` where this should apply. Also updates transpilers to use repeats - this actually occurs in `BobberBar`, probably not strictly necessary for all cases but shouldn't hurt either.

There is still some slight glitchiness when using the remapped button to finish the catch, i.e. dismiss the fish-caught dialog. Sometimes it doesn't work at all, sometimes it works and recasts, and sometimes (seen when the next fish was a jelly) it plays the animation of reeling in the next fish but doesn't actually catch it. Probably just need to detect this dialog state and suppress the activation.