Handle remapped button patch reversion and button release action separately.
We were previously using the `downButtons` to undo the patch, but since the logic had been (intentionally) changed to apply the patch even when not starting a continuous-activation cycle (i.e. due to `!Context.IsPlayerFree`), the patch wasn't always getting undone when the button was released.

The new logic checks both states separately, and undoes the patch if the button is physically up, regardless of whether it was tracked.

Also adds a custom activation action to end fishing, which seems to help with a few hard-to-identify edge cases where the default `pressToolButton` logic behaves oddly, either failing to clear the menu or doing a duplicate cast or "catching" the wrong fish/item.