Insert page for settings item when no pages are configured.
Initially the `ModMenuPages` may be empty; if the settings item is not explicitly disabled via config then we need to add an empty page containing just this one item.

Also adds a check for `ShowSettingsItem` in the case where there _are_ pages. This check was missing and causing the mod not to honor the disabled setting.