Postpone keybind-based activation from pre-update tick to post-update tick.
Most mods handle keybindings via an event like `ButtonsChanged`, or even their own `UpdateTicked`, thus "pressing" the buttons during the ticking event can cause vanilla behavior to be invoked before the target mod has a chance to catch and suppress it.

An example is the default AFS binding of Shift+F, which uses totally ordinary checks to watch for it (`ButtonsChanged` event plus `KeybindList.JustPressed`); firing this from the radial menu would open the Journal instead (shortcut "F") because the press simulation happens too early.

It is important not to actually modify any of the sequence/timing in the `MenuController` or the `UpdateTicking` logic that calls it, because menu state has to be carefully coordinated with player state. So we only delay the literal "key press" simulation, and still handle the "activation" as usual during pre-update.