Remap raw gamepad state onto actual input (SMAPI-wrapped) state instead of remapping raw state directly.
State changes don't always happen instantly, since SMAPI's input state is based on a snapshot, and we can confuse the game a little _too_ much (beyond the intentional trickery) into thinking that the X button was never actually released, or was released and then immediately pressed again the next frame.

Note that it's not sufficient to naively test whether the remapped button is actually "X" and skip remapping in that case. It does solve the specific bug, but subsequently prevents the rod from being used correctly when _explicitly_ bound to X, the result being an immediate cast without the ability to time/charge it.

This way works for both cases since it lets SMAPI input stay in control and injects additional buttons into its result, instead of replacing the entire state.

Fixes #3