Add workaround for unregistered item icons.
The pattern with these implementations seems to be a lack of object data and a custom implementation of `drawInMenu`. So, to reconstruct the texture, we can draw to a render target.

This is not ideal, could be subject to accidental cache poisoning or staleness, etc. But the alternative is just an error item sprite which is probably worse. There's also a performance hit, but we'll hope it's manageable for the relatively small size of a player inventory and the likelihood of these items being in it.

If texture derivation fails then it should fall back to the error icon it was displaying before.